defining and visualizing the ambulance domain
RandomBoolPolicy.java is the given random boolean policy code written in java. ambulance.rddl is the rddl code defining the ambulance domain. AmbulanceDisplay.java is a class inheriting 'StateViz' class, which defines components necessary for visualizing the domain and objects moving on it. 

Note that numbers of rows and columns used in AmbulanceDisplay.java should match with those used in defining an instance, e.g. ambulance_inst1.rddl.